updateHandbook functionality added
