+++ src/pages/agenda.js
@@ -1,106 +1,51 @@
-import React, { Component } from "react";
+import React, { useEffect, useState } from "react";
 import Collapsible from "react-collapsible";
-import { Placeholder } from "./updateHandbook";
-import Button from 'react-bootstrap/Button';
-import DropdownButton from 'react-bootstrap/DropdownButton';
-import Dropdown from 'react-bootstrap/Dropdown';
-import ButtonGroup from 'react-bootstrap/ButtonGroup';
-import { getSenateMeetingAllAPI, getSenatePointsAllAPI, getSenatePointsIdAPI, getSenatePointsMeetingIdAPI, addSenateMeetingAPI, addSenatePointAPI, } from '../api/senateMeeting'
 
-export function PointHistory(props) {
-    return (
-        <div className="point_history_tile">
-            <h5>Change in: {props.when}</h5>
-            <p>Modification: {props.change}</p>
-        </div>
-    );
-}
+import { getSenateMeetingAllAPI, getSenatePointsMeetingIdAPI, addSenateMeetingAPI, addSenatePointAPI } from '../api/senateMeeting'
+import Snackbar from "@mui/material/Snackbar";
+import Alert from "@mui/material/Alert";
 
 
 
-export function AgendaTile(props) {
-    return (
-        <li onClick={props.handleAgendaClick} className="list-item agenda" data-target={"agenda" + props.num}>{props.name}</li>
-    )
-}
+const Agenda = () => {
+    const [data, setdata] = useState();
+    const [agenda, setagenda] = useState(0);
+    const [point, setpoint] = useState(0);
+    const [pointData, setpointData] = useState();
+    const [NewPoint, setNewPoint] = useState(false);
+    const [number, setnumber] = useState();
+    const [name, setname] = useState();
 
-export function AgendaPointCreateForm(props) {
-    return (<div>
-
-        <label for="ag_pt_new_prop" class="form-label">Proposal</label>
-        <input type="number" class="form-control" id="ag_pt_new_prop" placeholder="Enter point number" onChange={props.handleChangenumber}></input>
-
-        <textarea type="text" class="form-control" id="ag_pt_new_prop" placeholder="Proposal for the Agenda Point" onChange={props.handleChangeProposal}></textarea>
-        {/* <input type="text" class="form-control" id="ag_pt_new_prop" placeholder="Resolution for the Agenda Point" onChange={props.handleChangeresolution}></input> */}
-
-        <div>
-            <button className="btn btn-success" onClick={props.handleClickCreate}>Create</button>
-            <button className="btn btn-danger" onClick={props.handleClickCancel}>Cancel</button>
-        </div>
-    </div>
-
-    )
-}
+    const [proposal, setproposal] = useState();
+    const [newpointCreated, setnewpointCreated] = useState(false);
 
 
-export function AgendaPointView(props) {
-    return (
-        <form className="ag_pt_view">
-            <label for="ag_pt_view_prop" class="form-label">Proposal</label>
-            <input readOnly={true} type="text" class="form-control" id="ag_pt_view_prop" placeholder={props.name} ></input>
-            <input readOnly={true} type="text" class="form-control" id="ag_pt_view_prop" placeholder={props.resolution}></input>
-            <button className="btn btn-danger">Delete Point</button>
-            {/*             
-            <div>
-                <button className="btn btn-success">Add Subpoint</button>
-                <button className="btn btn-primary">Edit</button>
-                <button className="btn btn-danger">Delete</button>
-            </div> */}
-        </form>
-    )
-}
-
-// export function
 
 
-class Agenda extends Component {
-    constructor(props) {
-        super(props);
-        this.state = { agenda: 0, point: 0, subpoint: 0, pointdata: {}, editing_point: false, newpoint: 0, proposal: '', number: 0, resolution: '', data: [] }
-        this.handleClickCreate = this.handleClickCreate.bind(this);
-        // this.state.data = [{num:3,name:'Agenda 5',points:[
-        //     {num:1,name:"Ag Pt 1", has_subpoints:true, subpoints:[
-        //         {num:1,name:"Subpoint 1"},
-        //         {num:2,name:"Subpoint 2"},
-        //         {num:3,name:"Subpoint 3"},
-        //     ]},
-        //     {num:1,name:"Ag Pt 2", has_subpoints:true, subpoints:[
 
-        //     ]},
-        //     {num:1,name:"Ag Pt 2", has_subpoints:false}
-        // ]},
-        // {num:2,name:'Agenda 2',points:[
-        //     {num:1,name:"Ag Pt 1", has_subpoints:true, subpoints:[
-        //         {num:1,name:"Subpoint 1"},
-        //         {num:2,name:"Subpoint 2"},
-        //     ]},
-        //     {num:1,name:"Ag Pt 2", has_subpoints:true, subpoints:[
-        //         {num:1,name:"Subpoint 1"},
-        //         {num:2,name:"Subpoint 2"},
-        //         {num:2,name:"Subpoint 3"},
-        //     ]},
-        // ]},]
+
+    const handleAgendaClick = (e) => {
+        setagenda(Number(e.target.dataset.target.split("_").slice(-1)));
+        setpoint(0);
     }
 
 
+    const handleNewPoint = async () => {
+        await addSenatePointAPI(number, name, proposal, agenda);
+        setnewpointCreated(true);
+        getdata();
+        setNewPoint(false)
 
 
-    async componentWillMount() {
+    }
 
+
+    const getdata = async () => {
         const response = await getSenateMeetingAllAPI();
-
         const Meeting = response.body;
         const new_data = []
+
+
 
         for (let i = 0; i < Meeting.length; i++) {
             const pointsArray = []
@@ -123,7 +68,9 @@ class Agenda extends Component {
                 // }
                 pointsArray.push({
                     num: points.number,
-                    name: points.proposal,
+                    name: points.name,
+                    proposal: points.proposal,
+                    resolution: points.resolution,
                     has_subpoints: false
                 })
             }
@@ -136,177 +83,145 @@ class Agenda extends Component {
             })
         };
 
-        this.setState({
-            data: new_data
-        }, this.forceUpdate())
-
-
+        console.log(new_data);
+        setdata(new_data)
 
     }
 
+    useEffect(() => {
+        getdata();
+    }, [])
+
+    return (
+        <>
 
 
-    handleAgendaClick = (e) => {
-        this.setState({
-            agenda: Number(e.target.dataset.target.split("_").slice(-1)),
-            point: 0
-        })
-    }
-
-    handleDeleteClick = (e) =>{
-
-    }
-
-    handleRenameClick = (e) =>{
-        
-    }
-
-    handleFinaliseClick = (e) =>{
-        
-    }
-
-    handleActionClick = (e) => {
-
-    }
-
-    handlePointClick = (e) => {
-
-        this.setState({
-            point: Number(e.target.dataset.target.split("_").slice(-1))
-        })
-    }
-
-    handleDeleteAgenda = (e) => {
-        
-    }
-
-    handleFinaliseAgenda = (e) => {
-
-    }
+            <Snackbar open={newpointCreated} autoHideDuration={6000} onClose={() => { setnewpointCreated(false); }}>
+                <Alert onClose={() => { setnewpointCreated(false); }} severity="success" sx={{ width: '100%' }}>
+                    New Point Created!
+                </Alert>
+            </Snackbar>
 
 
-    handleChangeProposal = (e) => {
-        this.setState({ proposal: e.target.value })
-    }
-
-    handleChangenumber = (e) => {
-        this.setState({ number: e.target.value })
-    }
-    handleChangeresolution = (e) => {
-        this.setState({ resolution: e.target.value })
-    }
-
-    async handleClickCreate(e) {
-
-        await addSenatePointAPI(this.state.number, this.state.proposal, this.state.resolution, this.state.agenda)
-        window.location.reload(false);
-
-    }
+            <div className="agenda-cont">
+                <div className="col-sm-3 agenda-menu left-pane">
+                    <button className="btn mb-3" data-target="create-agenda">Create New Agenda</button>
+                    <h2>Agendas</h2>
+                    <ul className="list-group">
+                        {
+                            data?.map((val) => {
+                                return <li data-active={agenda === val.num} onClick={handleAgendaClick} className="list-item agenda clickable" data-target={"agenda_" + val.num}>{val.name}</li>
+                            })
+                        }
+                    </ul>
+                </div>
 
 
-    handleClickCancel = () => {
-        this.setState({ newpoint: 0 })
-    }
 
-    handlenewpoint = () => {
-        this.setState({ newpoint: 1 })
-    }
-
-    render() {
-        return (<div className="agenda-cont">
-            <div className="col-sm-3 agenda-menu left-pane">
-                <button className="btn mb-3" data-target="create-agenda">Create New Agenda</button>
-                <h2>Agendas</h2>
-                <ul className="list-group">
-                    {
-                        this.state.data.map((val) => {
-                            return <li data-active={this.state.agenda === val.num} onClick={this.handleAgendaClick} className="list-item agenda clickable" data-target={"agenda_" + val.num}>{val.name}</li>
-                        })
-                    }
-                </ul>
-            </div>
-
-            <div className="col-sm-3 agenda-submenu center-pane">
-                {this.state.agenda !== 0 && <React.Fragment> 
-                    {/* <DropdownButton onSelect={(e) => console.log(e)} as={ButtonGroup} title="Dropdown" id="bg-nested-dropdown">
+                <div className="col-sm-3 agenda-submenu center-pane">
+                    {agenda !== 0 && <React.Fragment>
+                        {/* <DropdownButton onSelect={(e) => console.log(e)} as={ButtonGroup} title="Dropdown" id="bg-nested-dropdown">
                         <Dropdown.Item eventKey="1">Delete Selected Agenda</Dropdown.Item>
                         <Dropdown.Item eventKey="2">Finalise Agenda for Senate</Dropdown.Item>
                         <Dropdown.Item eventKey="3">Rename Selected Agenda</Dropdown.Item>
                     </DropdownButton> */}
-                    
-                    <button className="btn btn-primary mb-3" data-target="create-agenda" onClick={this.handlenewpoint}>New Agenda Point</button> 
-                <div className="btn-group">
-                    <button className="btn btn-danger mb-3" data-target="create-agenda" onClick={this.handleDeleteAgenda}>Delete Agenda</button> 
-                    <button className="btn btn-success mb-3" data-target="create-agenda" onClick={this.handleFinaliseAgenda}>Finalise Agenda</button>
-                </div>
-                <h2>Agenda Points</h2></React.Fragment>}
-                {this.state.data.map((val, id) => {
-                    return (
-                        this.state.agenda === val.num &&
-                        <ul id={"agenda_" + id} className="list-group">
-                            {val.points.map((det, id) => {
-                                return (
-                                    det.has_subpoints ?
-                                        (<Collapsible trigger={det.name}>
-                                            {
-                                                det.subpoints.map((sp) => {
-                                                    return (
-                                                        <li data-active={this.state.point === det.num} onClick={this.handlePointClick} className="list-item agenda-point clickable" data-target={"agenda-pt_" + sp.num}>Handbook Sub Point {sp.num}</li>
-                                                    )
-                                                })
-                                            }
-                                        </Collapsible>)
 
-                                        : <li data-active={this.state.point === det.num} onClick={this.handlePointClick} className="list-item agenda-point clickable" data-target={"agenda-pt_" + det.num}>{det.name}</li>
+                        <button className="btn btn-primary mb-3" data-target="create-agenda" onClick={() => { setNewPoint(true) }}>New Agenda Point</button>
+                        <button className="btn btn-danger mb-3" data-target="create-agenda" >Delete Agenda</button>
 
-                                )
-                            })}
-
-                        </ul>
-                    )
-                })
-                }
-
-            </div>
-
-            <div className="col-sm-6 changes-data pt-5">
-                {
-                    this.state.point === 0 ? (this.state.agenda === 0 ?
-                        <Placeholder text="agenda and point" feature="view/edit" /> :
-                        (this.state.newpoint === 1 ? <AgendaPointCreateForm handleChangeProposal={this.handleChangeProposal}
-                            handleChangeresolution={this.handleChangeresolution}
-                            handleChangenumber={this.handleChangenumber}
-                            handleClickCreate={this.handleClickCreate}
-                            handleClickCancel={this.handleClickCancel}
-
-                        /> : <Placeholder text="point" feature="view/edit" />)
-                    ) :
-                        // <AgendaPointView />
-                        null
-                }
-
-
-
-
-
-                {/* <pointHistory className="test">Hello</pointHistory> */}
-                {this.state.data?.map((val) => {
-                    return (
-                        val.points?.map((pt) => {
-                            return (
-                                this.state.agenda === val.num && this.state.point === pt.num &&
-                                pt.changes?.map((ch) => {
+                        <h2>Agenda Points</h2></React.Fragment>}
+                    {data?.map((val, id) => {
+                        return (
+                            agenda === val.num &&
+                            <ul id={"agenda_" + id} className="list-group">
+                                {val.points.map((det, id) => {
                                     return (
-                                        <PointHistory when={ch.when} change={ch.change}></PointHistory>
+                                        det.has_subpoints ?
+                                            (<Collapsible trigger={det.name}>
+                                                {
+                                                    det.subpoints.map((sp) => {
+                                                        return (
+                                                            <li data-active={point === det.num} onClick={this.handlePointClick} className="list-item agenda-point clickable" data-target={"agenda-pt_" + sp.num}>Handbook Sub Point {sp.num}</li>
+                                                        )
+                                                    })
+                                                }
+                                            </Collapsible>)
+
+                                            : <li data-active={point === det.num} onClick={(e) => {
+                                                setpoint(Number(e.target.dataset.target.split("_").slice(-1)))
+                                                setpointData(det);
+                                                console.log("agenda, point", agenda, point)
+                                                console.log("data", det)
+                                            }} className="list-item agenda-point clickable" data-target={"agenda-pt_" + det.num}>{det.name}</li>
+
                                     )
-                                })
+                                })}
+
+                            </ul>
+                        )
+                    })
+                    }
+
+                </div>
+
+
+
+
+                <div className="col-sm-6 changes-data pt-5">
+                    {
+                        NewPoint && (
+                            <div>
+
+                                <label for="ag_pt_new_prop" class="form-label">Proposal</label>
+                                <input type="number" class="form-control" id="ag_pt_new_prop" placeholder="Enter point number" onChange={(e) => { setnumber(e.target.value) }}></input>
+                                <input type="text" class="form-control" id="ag_pt_new_prop" placeholder="Enter point Name" onChange={(e) => { setname(e.target.value) }}></input>
+
+                                <textarea type="text" class="form-control" id="ag_pt_new_prop" placeholder="Proposal for the Agenda Point" onChange={(e) => { setproposal(e.target.value) }}></textarea>
+
+                                <div>
+                                    <button className="btn btn-success" onClick={handleNewPoint} >Create</button>
+                                    <button className="btn btn-danger" onClick={() => { setNewPoint(false) }}>Cancel</button>
+                                </div>
+                            </div>
+                        )
+                    }
+
+                    {
+                        !NewPoint && (point === 0 ? (agenda === 0 ?
+                            <h5 className="placeholder-tile">Select agenda and point first to view/edit</h5>
+                            : (
+                                <h5 className="placeholder-tile">Select point first to view/edit</h5>
                             )
-                        })
-                    )
-                })}
+                        ) : (<div>
+
+                            <form className="ag_pt_view">
+                                <label for="ag_pt_view_prop" class="form-label">Proposal</label>
+                                <input readOnly={true} type="text" class="form-control" id="ag_pt_view_prop" placeholder={pointData.name} ></input>
+                                <input readOnly={true} type="text" class="form-control" id="ag_pt_view_prop" placeholder={pointData.proposal}></input>
+                                <button className="btn btn-danger">Delete Point</button>
+
+                            </form>
+
+
+                        </div>))
+
+                    }
+
+
+
+                </div>
+
+
+
+
+
+
 
             </div>
-        </div>);
-    }
+        </>
+    )
+
+
 }
+
 export default Agenda;
